Refactoring

+++ src/Message/ConfirmMSG.cpp
@@ -73,7 +73,7 @@ namespace msg
 		Window dialogWindow(m_width, m_height, m_posX, m_posY);
 		dialogWindow.show();
 
-		ui::ConsoleManager::getInstance().setCursorPosition(m_width / 2 - m_title.size() / 2 + m_posX, m_posY + 1);
+		ui::ConsoleManager::getInstance().setCursorPosition(m_width / 2 - int(m_title.size() / 2) + m_posX, m_posY + 1);
 		std::cout << m_title;
 	}
