+++ src/Message/NotificationMSG.cpp
@@ -60,7 +60,7 @@ namespace msg
 		notificationWindow.addWindowName("notification", 1, 0);
 		notificationWindow.show();
 
-		ui::ConsoleManager::getInstance().setCursorPosition(m_width / 2 - m_title.size() / 2 + m_posX, m_posY + 1);
+		ui::ConsoleManager::getInstance().setCursorPosition(m_width / 2 - int(m_title.size() / 2) + m_posX, m_posY + 1);
 		std::cout << m_title;
 	}
